Switch back to modern rpm spec
The modern spec could be used to build rpm on supported Fedora
releases, it could not be used to build on epel8 and earlier.

For build rpm on epel8 and earlier, must use the old 201x era
spec which don't need exist in the upstream repo.

Remove the setuptools_scm dependency in pyproject.toml as it's
for source build not for packaging.

Signed-off-by: Wayne Sun <gsun@redhat.com>